Moves SchoolSelect into utilities dir
